requirements.txt
numpy==1.23.0
scikit-learn==1.1.0
streamlit==1.10.0
tensorflow-cpu
